chfsd: -s option for the database size (default 256 MiB) chfsd: -T option for the number of threads (default 5) chfsd: create a database directory if it does not exist
git-svn-id: svn+ssh://platina.ccs.tsukuba.ac.jp/svnroot/home/trunk/mochi/chfs@2488 7c857e19-e864-4221-8d65-6e103b07b3e5

+++ kv.h
@@ -1,4 +1,4 @@
-void kv_init(char *db_dir, char *engine, char *path, uint64_t size);
+void kv_init(char *db_dir, char *engine, char *path, size_t size);
 void kv_term();
 
 int kv_put(void *key, size_t key_size, void *value, size_t value_size);
